+++ R/ni_fleming.R
@@ -61,7 +61,7 @@ ni_fleming <- function(ve_lci, alpha = 0.025, power = 0.90){
     "Alpha" = alpha,
     "Power" = power,
     "Total number of events" = nsize,
-    "Max HR to declare NI" = as.numeric(max_hr),
+    "Max HR to declare NI" = unname(max_hr),
     "Max number of events in the experimental group" = max_i
   )
 }